+++ src/collective/workspace/membership.py
@@ -85,23 +85,33 @@ class TeamMembership(object):
             )
         super(TeamMembership, self).__setattr__(name, value)
 
-    def _update_groups(self, old_groups, new_groups, add_auto_groups=True):
+    def _update_groups(self, old_groups, new_groups):
         workspace = self.workspace
         context = workspace.context
         uid = context.UID()
         workspace_groups = get_workspace_groups_plugin()
-
         if self.user is None:
             return
 
+        for group_name in new_groups:
+            if group_name not in workspace.available_groups:
+                raise ValueError(
+                    'Unknown workspace group: {}'.format(group_name))
+        if self.groups != new_groups:
+            self.__dict__['groups'] = new_groups.copy()
+
         # Determine automatic groups
+        new_groups = set(new_groups)
+        old_groups = set(old_groups)
         for name, condition in workspace.auto_groups.items():
             if name not in workspace.available_groups:
                 raise Exception('Unknown workspace group: {}'.format(name))
-            if add_auto_groups and condition(self):
-                new_groups = new_groups.copy() | set([name])
+            # only add the automatic groups if condition is satisfied,
+            # otherwise remove it
+            if condition(self):
+                new_groups.add(name)
             else:
-                old_groups = old_groups.copy() | set([name])
+                old_groups.add(name)
 
         # Add to new groups
         for group_name in (new_groups - old_groups):
@@ -122,6 +132,18 @@ class TeamMembership(object):
             except KeyError:  # group doesn't exist
                 pass
 
+    def _remove_all_groups(self):
+        workspace = self.workspace
+        uid = workspace.context.UID()
+        workspace_groups = get_workspace_groups_plugin()
+        groups = set(self.groups) | set(workspace.available_groups)
+        for group_name in groups:
+            group_id = '{}:{}'.format(group_name.encode('utf8'), uid)
+            try:
+                workspace_groups.removePrincipalFromGroup(self.user, group_id)
+            except KeyError:  # group doesn't exist
+                pass
+
     @property
     def groups(self):
         # Don't include automatic groups
@@ -136,7 +158,7 @@ class TeamMembership(object):
         if 'user' in data and old['user'] != data['user']:
             # User is changing, so remove the old user from groups.
             user_changed = True
-            self._update_groups(old['groups'], set())
+            self._remove_all_groups()
         data['_mtime'] = DateTime()
         self.__dict__.update(data)
         # make sure change is persisted
@@ -187,7 +209,7 @@ class TeamMembership(object):
             if func(self.__dict__):
                 workspace.context._counters[name].change(-1)
         del self.workspace.members[self._key]
-        self._update_groups(self.groups, set(), add_auto_groups=False)
+        self._remove_all_groups()
         self.handle_removed()
         notify(TeamMemberRemovedEvent(self.workspace.context, self))
         self.workspace.context.reindexObject(idxs=['workspace_members'])